// src/app/models/product.ts
export interface Product {
  key: string;
  ballaststoffe: number;
  ballaststoffePro: number;
  category: string;
  davonGes: number;
  davonGesPro: number;
  davonZucker: number;
  davonZuckerPro: number;
  eiweiss: number;
  eiweissPro: number;
  energieKJ: number;
  energieKJPro: number;
  energieKcal: number;
  energieKcalPro: number;
  fettJe: number;
  fettPro: number;
  imageUrl: string;
  kohlenhydrate: number;
  kohlenhydratePro: number;
  pictureNum: number;
  salzJe: number;
  salzPro: number;
  title: string;
  volume: number;
}